fix($point): 포인트 타입 Long으로 변경
point, amount 타입 Long으로 통일

+++ src/main/java/comatching/comatching3/users/controller/UserController.java
@@ -91,7 +91,7 @@ public class UserController {
 	 */
 	@GetMapping("/auth/user/api/points")
 	public Response<String> getPoints() {
-		Integer points = userService.getPoints();
+		Long points = userService.getPoints();
 		return Response.ok("point : " + points);
 	}
